refactor: 拆分2048小游戏hooks引擎
