+++ src/renderer/components/states/maxExpand/components/games/2048/utils/position.ts
@@ -1,0 +1,8 @@
+import { CELL, GAP, PAD } from '../config/constants';
+
+export function resolveTilePosition(row: number, col: number): { left: number; top: number } {
+  return {
+    left: PAD + col * (CELL + GAP),
+    top: PAD + row * (CELL + GAP),
+  };
+}
